gui close to being done. need labels + to add hooks. Dragging after.

+++ source/planet.cpp
@@ -6,10 +6,10 @@
 
 Planet::Planet(const double mass, const double radius, const sf::Vector2f velocity, const sf::Vector2f position, const sf::Color color, const bool isStatic)
 : mass(mass), radius(radius), velocity(velocity), position(position), color(color), isStatic(isStatic){
-    shape.setRadius(radius * RADIUS_SCALE);
+    shape.setRadius(static_cast<float>(radius * RADIUS_SCALE));
     shape.setFillColor(color);
     shape.setOrigin(shape.getRadius(), shape.getRadius());
-    shape.setPosition(position.x * DISTANCE_SCALE, position.y * DISTANCE_SCALE);
+    shape.setPosition(static_cast<float>(position.x * DISTANCE_SCALE), static_cast<float>(position.y * DISTANCE_SCALE));
 }
 
 bool PlanetGroup::addPlanet(const std::string& name, const Planet& p) {
